requirements.txt
torch>=1.7
torchvision>=0.8
albumentations
opencv-python
matplotlib
numpy
pillow